Add new Storage class
