+++ src/main/java/TaskList.java
@@ -1,0 +1,2 @@
+public class TaskList {
+}
